finish basic static-blog support

+++ src/AcBlog.Server.API/Controllers/PostsController.cs
@@ -1,8 +1,9 @@
 ﻿using System.Collections.Generic;
+using System.Net.NetworkInformation;
 using System.Threading.Tasks;
 using AcBlog.Data.Models;
 using AcBlog.Data.Models.Actions;
-using AcBlog.Data.Providers;
+using AcBlog.Data.Repositories;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -12,25 +13,56 @@ namespace AcBlog.Server.API.Controllers
     [ApiController]
     public class PostsController : ControllerBase
     {
-        IPostProvider Provider { get; }
+        IPostRepository Provider { get; }
 
-        public PostsController(IPostProvider provider)
+        public PostsController(IPostRepository provider)
         {
             Provider = provider;
         }
 
+        [HttpGet("actions/read")]
+        public async Task<ActionResult<bool>> CanRead()
+        {
+            return await Provider.CanRead();
+        }
+
+        [HttpGet("actions/write")]
+        public async Task<ActionResult<bool>> CanWrite()
+        {
+            return await Provider.CanWrite();
+        }
+
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Post>>> All() => Ok(await Provider.All());
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesDefaultResponseType]
+        public async Task<ActionResult<IEnumerable<Post>>> All()
+        {
+            if (!await Provider.CanRead())
+                return BadRequest();
+            return Ok(await Provider.All());
+        }
 
         [HttpPut("query")]
-        public async Task<ActionResult<PostQueryResponse>> Query([FromBody] PostQueryRequest query) => Ok(await Provider.Query(query));
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesDefaultResponseType]
+        public async Task<ActionResult<QueryResponse<string>>> Query([FromBody] PostQueryRequest query)
+        {
+            if (!await Provider.CanRead())
+                return BadRequest();
+            return Ok(await Provider.Query(query));
+        }
 
         [HttpGet("{id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesDefaultResponseType]
         public async Task<ActionResult<Post>> Get(string id)
         {
+            if (!await Provider.CanRead())
+                return BadRequest();
             if (await Provider.Exists(id))
                 return Ok(await Provider.Get(id));
             else
@@ -39,19 +71,25 @@ namespace AcBlog.Server.API.Controllers
 
         [HttpPost]
         [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesDefaultResponseType]
         public async Task<ActionResult<string>> Create([FromBody] Post value)
         {
+            if (!await Provider.CanWrite())
+                return BadRequest();
             var result = await Provider.Create(value);
             return Created(result, result);
         }
 
         [HttpPut("{id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesDefaultResponseType]
         public async Task<ActionResult<bool>> Update(string id, [FromBody] Post value)
         {
+            if (!await Provider.CanWrite())
+                return BadRequest();
             value.Id = id;
             if (await Provider.Exists(value.Id))
                 return Ok(await Provider.Update(value));
@@ -61,10 +99,13 @@ namespace AcBlog.Server.API.Controllers
 
         [HttpDelete("{id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesDefaultResponseType]
         public async Task<ActionResult<bool>> Delete(string id)
         {
+            if (!await Provider.CanWrite())
+                return BadRequest();
             if (await Provider.Exists(id))
                 return Ok(await Provider.Delete(id));
             else
