+++ src/AcBlog.Server.API/Controllers/UsersController.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using AcBlog.Data.Models;
-using AcBlog.Data.Providers;
+using AcBlog.Data.Models.Actions;
+using AcBlog.Data.Repositories;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -13,22 +14,56 @@ namespace AcBlog.Server.API.Controllers
     [ApiController]
     public class UsersController : ControllerBase
     {
-        IUserProvider Provider { get; }
+        IUserRepository Provider { get; }
 
-        public UsersController(IUserProvider provider)
+        public UsersController(IUserRepository provider)
         {
             Provider = provider;
         }
 
+        [HttpGet("actions/read")]
+        public async Task<ActionResult<bool>> CanRead()
+        {
+            return await Provider.CanRead();
+        }
+
+        [HttpGet("actions/write")]
+        public async Task<ActionResult<bool>> CanWrite()
+        {
+            return await Provider.CanWrite();
+        }
+
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<User>>> All() => Ok(await Provider.All());
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesDefaultResponseType]
+        public async Task<ActionResult<IEnumerable<User>>> All()
+        {
+            if (!await Provider.CanRead())
+                return BadRequest();
+            return Ok(await Provider.All());
+        }
+
+        [HttpPut("query")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesDefaultResponseType]
+        public async Task<ActionResult<QueryResponse<string>>> Query([FromBody] UserQueryRequest query)
+        {
+            if (!await Provider.CanRead())
+                return BadRequest();
+            return Ok(await Provider.Query(query));
+        }
 
         [HttpGet("{id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesDefaultResponseType]
         public async Task<ActionResult<User>> Get(string id)
         {
+            if (!await Provider.CanRead())
+                return BadRequest();
             if (await Provider.Exists(id))
                 return Ok(await Provider.Get(id));
             else
@@ -37,19 +72,25 @@ namespace AcBlog.Server.API.Controllers
 
         [HttpPost]
         [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesDefaultResponseType]
         public async Task<ActionResult<string>> Create([FromBody] User value)
         {
+            if (!await Provider.CanWrite())
+                return BadRequest();
             var result = await Provider.Create(value);
             return Created(result, result);
         }
 
         [HttpPut("{id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesDefaultResponseType]
         public async Task<ActionResult<bool>> Update(string id, [FromBody] User value)
         {
+            if (!await Provider.CanWrite())
+                return BadRequest();
             value.Id = id;
             if (await Provider.Exists(value.Id))
                 return Ok(await Provider.Update(value));
@@ -59,10 +100,13 @@ namespace AcBlog.Server.API.Controllers
 
         [HttpDelete("{id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesDefaultResponseType]
         public async Task<ActionResult<bool>> Delete(string id)
         {
+            if (!await Provider.CanWrite())
+                return BadRequest();
             if (await Provider.Exists(id))
                 return Ok(await Provider.Delete(id));
             else
